requirements.txt
fastapi
uvicorn
python-multipart
huggingface_hub
python-dotenv
Pillow